Sliding Door
Added a functioning door in Alan's Test Scene that slides up and down with no current animation. Created a new lever variant called "Door Lever Variant". New script titled "Door" that allows for the object it is placed on to slide up and down when the proper lever variant is pulled.